+++ node/clients/Oms.ts
@@ -8,7 +8,7 @@ export class OMSClient extends JanusClient {
     super(ctx, {
       ...options,
       headers: {
-        ...(options && options.headers),
+        ...options?.headers,
         VtexIdclientAutCookie: ctx.authToken,
       },
     })