+++ node/resolvers/index.ts
@@ -17,66 +17,15 @@ import {
 } from '../mdSchema'
 import { toHash } from '../utils'
 import GraphQLError from '../utils/GraphQLError'
+import { organizationName, costCenterName } from './fieldResolvers'
 
 const getAppId = (): string => {
   return process.env.VTEX_APP_ID ?? ''
 }
 
-interface OrganizationInput {
-  name: string
-  b2bCustomerAdmin: B2BCustomerInput
-  defaultCostCenter: DefaultCostCenterInput
-}
-
-interface B2BCustomerInput {
-  firstName: string
-  lastName: string
-  email: string
-}
-
-interface DefaultCostCenterInput {
-  name: string
-  address: AddressInput
-}
-
-interface CostCenterInput {
-  name: string
-  addresses: [AddressInput]
-}
-
 const CONNECTOR = {
   PROMISSORY: 'Vtex.PaymentGateway.Connectors.PromissoryConnector',
 } as const
-
-interface AddressInput {
-  addressId: string
-  addressType: string
-  postalCode: string
-  country: string
-  receiverName: string
-  city: string
-  state: string
-  street: string
-  number: string
-  complement: string
-  neighborhood: string
-  geoCoordinates: [number]
-}
-
-interface OrganizationRequest {
-  name: string
-  defaultCostCenter: DefaultCostCenterInput
-  b2bCustomerAdmin: string
-  status: string
-  created: string
-}
-
-interface Organization {
-  name: string
-  costCenters: string[]
-  status: string
-  created: string
-}
 
 const defaultSettings = {
   adminSetup: {
@@ -160,6 +109,36 @@ const QUERIES = {
         slug
       }
       permissions
+    }
+  }`,
+  listUsers: `query users($organizationId: ID, $costCenterId: ID) {
+    listUsers(organizationId: $organizationId, costCenterId: $costCenterId) @context(provider: "vtex.storefront-permissions") {
+      id
+      roleId
+      userId
+      clId
+      orgId
+      costId
+      name
+      email
+      canImpersonate
+    }
+  }`,
+}
+
+const MUTATIONS = {
+  saveUser: `mutation saveUser($id: ID, $roleId: ID!, $userId: ID, $orgId: ID, $costId: ID, $clId: ID, $canImpersonate: Boolean, $name: String!, $email: String!) {
+    saveUser(id: $id, roleId: $roleId, userId: $userId, orgId: $orgId, costId: $costId, clId: $clId, canImpersonate: $canImpersonate, name: $name, email: $email) @context(provider: "vtex.storefront-permissions") {
+      id
+      status
+      message
+    }
+  }`,
+  deleteUser: `mutation deleteUser($id: ID!, $userId: ID, $email: String!) {
+    deleteUser(id: $id, userId: $userId, email: $email) @context(provider: "vtex.storefront-permissions") {
+      id
+      status
+      message
     }
   }`,
 }
@@ -568,11 +547,27 @@ export const resolvers = {
     ) => {
       const {
         clients: { masterdata },
+        vtex,
         vtex: { logger },
       } = ctx
 
       // create schema if it doesn't exist
       await checkConfig(ctx)
+
+      if (!organizationId) {
+        // get user's organization from session
+        const { sessionData } = vtex as any
+
+        if (!sessionData?.namespaces['storefront-permissions']) {
+          throw new GraphQLError('organization-data-not-found')
+        }
+
+        const {
+          organization: { value: userOrganizationId },
+        } = sessionData.namespaces['storefront-permissions']
+
+        organizationId = userOrganizationId
+      }
 
       try {
         const costCenter = {
@@ -717,7 +712,7 @@ export const resolvers = {
         clients: { masterdata },
       } = ctx
 
-      // TODO: also delete organization's cost centers
+      // TODO: also delete organization's cost centers?
 
       try {
         await masterdata.deleteDocument({
@@ -742,6 +737,7 @@ export const resolvers = {
       } = ctx
 
       // TODO: remove cost center from organization
+      // or just remove the costCenters array from the organization entity, probably don't need it
 
       try {
         await masterdata.deleteDocument({
@@ -759,6 +755,133 @@ export const resolvers = {
           throw new GraphQLError(e)
         }
       }
+    },
+    saveUser: async (
+      _: any,
+      {
+        id,
+        roleId,
+        userId,
+        orgId,
+        costId,
+        clId,
+        canImpersonate = false,
+        name,
+        email,
+      }: UserArgs,
+      ctx: Context
+    ) => {
+      const {
+        clients: { graphQLServer },
+        vtex,
+        vtex: { adminUserAuthToken, logger },
+      } = ctx
+
+      // create schema if it doesn't exist
+      await checkConfig(ctx)
+
+      const { sessionData, storefrontPermissions } = vtex as any
+
+      if (
+        !adminUserAuthToken &&
+        !sessionData?.namespaces['storefront-permissions']?.organization
+      ) {
+        throw new GraphQLError('organization-data-not-found')
+      }
+
+      if (
+        !adminUserAuthToken &&
+        !storefrontPermissions?.permissions?.includes('add-users-organization')
+      ) {
+        throw new GraphQLError('operation-not-permitted')
+      }
+
+      if (
+        !adminUserAuthToken &&
+        sessionData.namespaces['storefront-permissions'].organization !== orgId
+      ) {
+        throw new GraphQLError('operation-not-permitted')
+      }
+
+      const addUserResult = await graphQLServer
+        .mutation(MUTATIONS.saveUser, {
+          id,
+          roleId,
+          userId,
+          orgId,
+          costId,
+          clId,
+          canImpersonate,
+          name,
+          email,
+        })
+        .then((result: any) => {
+          return result.data.saveUser
+        })
+        .catch((error: any) => {
+          console.error(error)
+          logger.error({
+            message: 'addUser-error',
+            error,
+          })
+
+          return { status: 'error', message: error }
+        })
+
+      return addUserResult
+    },
+    removeUser: async (
+      _: any,
+      { id, userId, email }: UserArgs,
+      ctx: Context
+    ) => {
+      const {
+        clients: { graphQLServer },
+        vtex,
+        vtex: { adminUserAuthToken, logger },
+      } = ctx
+
+      // create schema if it doesn't exist
+      await checkConfig(ctx)
+
+      const { sessionData, storefrontPermissions } = vtex as any
+
+      if (
+        !adminUserAuthToken &&
+        !sessionData?.namespaces['storefront-permissions']?.organization
+      ) {
+        throw new GraphQLError('organization-data-not-found')
+      }
+
+      if (
+        !adminUserAuthToken &&
+        !storefrontPermissions?.permissions?.includes(
+          'remove-users-organization'
+        )
+      ) {
+        throw new GraphQLError('operation-not-permitted')
+      }
+
+      const deleteUserResult = await graphQLServer
+        .mutation(MUTATIONS.deleteUser, {
+          id,
+          userId,
+          email,
+        })
+        .then((result: any) => {
+          return result.data.deleteUser
+        })
+        .catch((error: any) => {
+          console.error(error)
+          logger.error({
+            message: 'deleteUser-error',
+            error,
+          })
+
+          return { status: 'error', message: error }
+        })
+
+      return deleteUserResult
     },
     saveAppSettings: async (_: any, __: any, ctx: Context) => {
       const {
@@ -833,7 +956,7 @@ export const resolvers = {
             schema: ORGANIZATION_REQUEST_SCHEMA_VERSION,
             pagination: { page, pageSize },
             sort: `${sortedBy} ${sortOrder}`,
-            ...(where ? { where } : {}),
+            ...(where && { where }),
           }
         )
 
@@ -936,7 +1059,7 @@ export const resolvers = {
             schema: ORGANIZATION_SCHEMA_VERSION,
             pagination: { page, pageSize },
             sort: `${sortedBy} ${sortOrder}`,
-            ...(where ? { where } : {}),
+            ...(where && { where }),
           }
         )
 
@@ -962,10 +1085,39 @@ export const resolvers = {
     ) => {
       const {
         clients: { masterdata },
+        vtex,
+        vtex: { adminUserAuthToken },
       } = ctx
 
       // create schema if it doesn't exist
       await checkConfig(ctx)
+
+      const { sessionData } = vtex as any
+
+      if (!id) {
+        // get user's organization from session
+        if (!sessionData?.namespaces['storefront-permissions']) {
+          throw new GraphQLError('organization-data-not-found')
+        }
+
+        const {
+          organization: { value: userOrganizationId },
+        } = sessionData.namespaces['storefront-permissions']
+
+        id = userOrganizationId
+      }
+
+      if (!adminUserAuthToken) {
+        // if not admin user, verify user has access to requested organization
+
+        const {
+          organization: { value: userOrganizationId },
+        } = sessionData.namespaces['storefront-permissions']
+
+        if (id !== userOrganizationId) {
+          throw new GraphQLError('operation-not-permitted')
+        }
+      }
 
       try {
         const organization = await masterdata.getDocument({
@@ -1062,10 +1214,40 @@ export const resolvers = {
     ) => {
       const {
         clients: { masterdata },
+        vtex,
+        vtex: { adminUserAuthToken },
       } = ctx
 
       // create schema if it doesn't exist
       await checkConfig(ctx)
+
+      const { sessionData } = vtex as any
+
+      if (!id) {
+        // get user's organization from session
+
+        if (!sessionData?.namespaces['storefront-permissions']) {
+          throw new GraphQLError('organization-data-not-found')
+        }
+
+        const {
+          organization: { value: userOrganizationId },
+        } = sessionData.namespaces['storefront-permissions']
+
+        id = userOrganizationId
+      }
+
+      if (!adminUserAuthToken) {
+        // if not admin user, verify user has access to requested organization
+
+        const {
+          organization: { value: userOrganizationId },
+        } = sessionData.namespaces['storefront-permissions']
+
+        if (id !== userOrganizationId) {
+          throw new GraphQLError('operation-not-permitted')
+        }
+      }
 
       let where = `organization=${id}`
 
@@ -1080,7 +1262,7 @@ export const resolvers = {
           schema: COST_CENTER_SCHEMA_VERSION,
           pagination: { page, pageSize },
           sort: `${sortedBy} ${sortOrder}`,
-          ...(where ? { where } : {}),
+          ...(where && { where }),
         })
 
         return costCenters
@@ -1097,19 +1279,39 @@ export const resolvers = {
     getCostCenterById: async (_: any, { id }: { id: string }, ctx: Context) => {
       const {
         clients: { masterdata },
+        vtex,
+        vtex: { adminUserAuthToken },
       } = ctx
 
       // create schema if it doesn't exist
       await checkConfig(ctx)
 
       try {
-        const organization = await masterdata.getDocument({
+        const costCenter: CostCenter = await masterdata.getDocument({
           dataEntity: COST_CENTER_DATA_ENTITY,
           fields: COST_CENTER_FIELDS,
           id,
         })
 
-        return organization
+        if (!adminUserAuthToken) {
+          // if not admin user, verify user has access to requested organization
+
+          const { sessionData } = vtex as any
+
+          if (!sessionData?.namespaces['storefront-permissions']) {
+            throw new GraphQLError('organization-data-not-found')
+          }
+
+          const {
+            organization: { value: userOrganizationId },
+          } = sessionData.namespaces['storefront-permissions']
+
+          if (costCenter.organization !== userOrganizationId) {
+            throw new GraphQLError('operation-not-permitted')
+          }
+        }
+
+        return costCenter
       } catch (e) {
         if (e.message) {
           throw new GraphQLError(e.message)
@@ -1142,6 +1344,50 @@ export const resolvers = {
           throw new GraphQLError(e)
         }
       }
+    },
+    getUsers: async (
+      _: any,
+      {
+        organizationId,
+        costCenterId,
+      }: { organizationId: string; costCenterId: string },
+      ctx: Context
+    ) => {
+      const {
+        clients: { graphQLServer },
+        vtex: { logger },
+      } = ctx
+
+      const variables = {
+        ...(organizationId && { organizationId }),
+        ...(costCenterId && { costCenterId }),
+      }
+
+      const users = await graphQLServer
+        .query(QUERIES.listUsers, variables, {
+          persistedQuery: {
+            provider: 'vtex.storefront-permissions@1.x',
+            sender: 'vtex.b2b-organizations@0.x',
+          },
+        })
+        .then((result: any) => {
+          return result.data.listUsers
+        })
+        .catch(error => {
+          logger.error({
+            message: 'getUsers-error',
+            error,
+          })
+          if (error.message) {
+            throw new GraphQLError(error.message)
+          } else if (error.response?.data?.message) {
+            throw new GraphQLError(error.response.data.message)
+          } else {
+            throw new GraphQLError(error)
+          }
+        })
+
+      return users
     },
     getAppSettings: async (_: any, __: any, ctx: Context) => {
       const {
@@ -1197,5 +1443,9 @@ export const resolvers = {
 
       return settings
     },
+  },
+  B2BUser: {
+    organizationName,
+    costCenterName,
   },
 }